refactor: centralize today helper
